backend/requirements.txt
# Insurance AI Demo - Backend API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.40.0,<2
httpx>=0.27.0,<0.28
azure-identity>=1.17.0
websockets>=12.0
aiohttp>=3.9.0
# Persistencia y autenticación
azure-cosmos>=4.7.0
PyJWT[crypto]>=2.9.0
